+++ Data/DropPool.cs
@@ -18,6 +18,8 @@ namespace CUCoreLib.Data
         Trader3 = 1 << 6,
         AllTraders = Trader1 | Trader2 | Trader3,
         DropCapsule = 1 << 7,
-        CapsuleContainer = 1 << 8
+        CapsuleContainer = 1 << 8,
+
+        All = Corpse | MedicalCrate | FoodCrate | ContainerCrate | AllTraders | DropCapsule | CapsuleContainer
     }
 }